// admin-clients/libs/cpanel/venues/feature/standard-venue-tpl-manager/src/lib/models/venue-tpl-tree-dialog-type.enum.ts
export enum SectorActionType {
    create = 'create',
    clone = 'clone',
    editName = 'editName',
    increaseCreate = 'increaseCreate',
    increaseEdit = 'increaseEdit'
}

export enum ZoneActionType {
    create = 'create',
    clone = 'clone',
    editName = 'editName',
    editCapacity = 'editCapacity',
    increaseCreate = 'increaseCreate',
    increaseEditCapacity = 'increaseEditCapacity'
}